Revert "refactor(go/adbc/driver): moved validation tests closer to drivers to have access to internal util functions"
This reverts commit d6d6a2706f251707675b83285d32467d7358b23c.
